Exit on SIGTERM
Akaros's select is a little nasty, and it won't wake up the thread if the
process receives a SIGTERM.  Actually, even if we knew which thread to send an
intra-process signal too, that wouldn't work either.  The uthread is blocked in
epoll_wait on two ev_qs, and doesn't expect to wake early.  We could build
that, but we still wouldn't know which thread in DB to interrupt.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>